Server/EncodingServer/src/main/java/com/visiblesound/gcp/upload_from_server_to_GCP.py
from google.cloud import storage

import sys

input_file_path = sys.argv[1]
output_file_path = sys.argv[2]


def upload_blob(bucket_name, source_file_name, destination_blob_name):
    """Uploads a file to the bucket."""
    storage_client = storage.Client()
    bucket = storage_client.get_bucket(bucket_name)
    blob = bucket.blob(destination_blob_name)

    blob.upload_from_filename(source_file_name)

    print('File {} uploaded to {}.'.format(
        source_file_name,
        destination_blob_name))

upload_blob("visible_voice",input_file_path,output_file_path)
