Server/EncodingServer/requirements.txt
argparse
google-cloud-speech
google-cloud-storage
wordcloud
matplotlib
pandas
